add: product service

+++ src/main/java/pl/bartoszsredzinski/ecommerceshopv1/repositoreis/CategoryCrudRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import pl.bartoszsredzinski.ecommerceshopv1.model.Category;
 
 @Repository
-public interface CategoryCrudRepository extends CrudRepository<Category, Long>{
+public interface CategoryCrudRepository extends CrudRepository<Category, Integer>{
     Category findByName(String name);
 }
